K3Process -> KProcess, ChecksumDialog
The checksum tools are now called directly (and not via a shell as it used
to be). Output redirection is now handled by QProcess directly which should
be much more portable. In order to remain the old behaviour I used the
blocking function QProcess::waitForStarted, since KProcess::start() is now
non-blocking and returns void. Here is still room for improvements so that
the tools are called compleatly asynchronously, see TODO-comments. This
however requires architectural changes on the checksum modul and may also
krusaderapp because of wasWaitingCancelled