FIXED: Filter: Don't list comments in user/group chooser
At least on Mac OS X /etc/passwd contains comments starting with '#'. These
comment lines caused the seachdialog to be way too large for the screen. If
there are platforms which allow the #-sign in user or groupnames this change
can be #ifdef'ed for __APPLE__ but I don't think it's nesseary.