FIXED: Actionman, sometimes the command field is empty, patch by Vaclav Juza, thanks
Crazy bug, it happened to me only for very few combinations
(for example when changing from Unmount to Mount), depending on lengths of
both commands (the first longer, but the second quite long as well). When I
clicked to that field, the text appeared. It seems, that the text is initially
scrolled out of the visible area. Probably a KDE or Qt bug.

It could be worked around by the attached patch (Puts cursor at the
beginning). Vaclav