FIXED: Useractions, after pwd command current workdir was not changed,            thanks to Vaclav Juza
Comment by Vaclav Juza:
When I added the pwd command before the actual useraction command, I realized
that it was because sometimes the current working directory was not changed
to the path of the panel.
I had packed in the right panel, then moved the archive to the left and
unpacked there. When I gave the focus to the left using the mouse click,
everything was fine, but when I changed the panel to the left using the
tab key, the working directory was not changed.