K3Process -> KProcess: 7z encryption checker
Because there is no shell any more we have to use a process group to work around
the broken 7z wrapper. Unfortunately QProcess::kill is not virtual so we have to
hack around this by calling the posix function kill() directly with the negative
pid to kill the whole group.