forward port KDE3 -> KDE4: FIXED: Current working directory not restored in populateVfsList
cvs reference:
revision 1.70
date: 2008/03/06 20:45:44;  author: codeknight;  state: Exp;  lines: +2 -0
So my last patch seems to have solved only some symptoms, but not the problem (/Panel/listpanel.cpp).
This patch really solves the problem
(when refreshing the panel, current working dir was changed to read the
contents of the directory and was not changed back).
My previous patch (/Panel/listpanel.cpp) is no more necessary (or useful). Vaclav