FIXED: [ 310722 ] viewer crash when opening file BUG: 310722
The crash is caused by qApp->processEvents() being called in PanelViewerBase::readFileInfo()
and the calling object being deleted before processEvents() returns.

Fix: remove call to qApp->processEvents() and change the control flow of opening files accordingly.

This necessitates some restructuring.

The logic determining which widget the focus should be returned to,
after the viewer is closed, is simplified:
The focus is always returned to the last widget
which was passed as 'parent' to KrViewer::view()/edit()
if it isn't NULL, otherwise to the main window.

KrViewer::createGUI()/addTab():
    handle the case of the tab not containing a KPart

make sure that:
    - event filters aren't installed multiple times
    - signals of the KPart aren't connected to multiple times

remove dead code