Don't allow implicit QString to QUrl conversion
Many APIs had changed from accepting parameters of type KUrl to QUrl. If the caller was passing a QString there was now a behaviour change as a result since QUrl interprets a string differently than KUrl does. So this fixes those porting regressions by not allowing this implicit conversion (adds define of QT_NO_URL_CAST_FROM_STRING) and then fixes all current such usages to correctly pass a QUrl instead.
Also, it turns out that the easiest way to fix the problem of QStrings passed to KFileDialog methods was to port the calls to QFileDialog, so do that as well (very straight forward porting anyway).

One more thing to note, added a static method to vfs class for ensuring url path ends with a slash, that is already needed in many places and will be needed even more when porting from KUrl to QUrl.