Port remaining parts from KUrl to QUrl
Mostly straight forward porting, using porting script and also manual labour. Either way, it follows the guidelines (doc comments) in kurl.h from KF5.
Mostly the difficult thing is the change in constructor that takes a QString. Needed to decide for each case if KUrl(QString) construction needed to be replaces with QUrl::fromUserInput or the simpler one, QUrl::fromLocalFile. If there are any regressions that would be likely candidate.
Also some additional changes that now made sense to do:
- KChooseDir dialogs: code could be simplified quite a lot, KUrlRequester does everything we needed already.
- Krusader class, constructor: making sure relative paths are resolved is now done by QUrl anyway, could remove code for that.