Fix broken application focus detection
The focusOutEvent and focusInEvent in KrusaderApp class did not override any virtual method (QApplication does not inherit QWidget) so they were never called.
Instead, hook up the applicationStateChanged signal from QGuiApplication directly to KRslots.
And since the original purpose of KrusaderApp class is now gone, remove it and move the remaining part of it (application startup detection) to the Krusader class.