Change API for KrAction name
The comment in the emtpy implementation of the setName method suggests that it was overriding some virtual method. But there is no such method. Switch to storing the name in the objectName property from QObject. That way we could even add a whole list of KrAction objects to a KActionCollection and it would work!