Fix broken Search dialog
Got broken from porting KUrl to QUrl, did not return any search results. Turns out the KUrlCompletion::replacedPath is a bit stupid, returns either a url or a path depending on if it's a local file or not.
So we need to have code that figures out if it was stripped of it's protocol or not.