Port from KInputDialog to QInputDialog
Mostly easy, one difference is that there is no way to ensure Ok button is only enabled when the QLineEdit contains text. KInputDialog did that. But I consider that a small and not important difference, now the user can hit enter without entering a text and the result will most likely be like pressing cancel, the receiving code checks if the text entered is empty and does nothing in that case.