Remove "minimize to system tray" feature.
Porting this to KF5 would mean rewriting the feature using KStatusNotifier instead. This commit can be seen as a first step, it removes the existing and deprecated implementation (which was not working on Plasma 5 anyway, XEmbed systray icons are not supported there). I myself have no plans of writing a new implementation for this feature.