Move Krusader::supportedTools to KrServices
This is an effort to untangle the big criss cross of includes across modules. So it changes some other things also. The idea was to try to reduce the number of frameworks each module needs to mention in it's target_link_libraries (in CMakeLists.txt...). Also gives benefit of not having to recompile so many other modules after changing a header file. There is still much to do, this was just a start... things that I came across today.