Change VFS API to make it safer
Returning a raw pointer to manually allocated memory from a function is bound to cause leaks.. And it did in this case. (Users of vfs_getFiles did not always free the result when done with it)
Instead, use QList as return type. QList object itself is basically just a ref counting smart pointer.

Also, this change kind of spread itself so also adapt vfs_delFiles and vfs_addFiles to take references instead of pointers.