Fix copying folders out from archive files, in a different way
Alternative fix for commit 60cf3b1d60c036daebdcd44fd8c87a1d670888d6, fix for copy/move folders from archive.

Change two implementations of vfs_getFile() to not return a url with a slash appended for folders. The normal_vfs already did not do this. And at least one user did not work with this.
To make things consistent the other option would be to make normal_vfs also append slash for folders. And then fix all users that did not want the slash to remove it again.
I still don't know if there is any user that needs the slash... it can't be any common operation at least, would have noticed the breakage easily since normal_vfs is the most commonly used.
The commit that introduced adding a slash mentioned that it was for fixing "following link in sftp", I have tested copying a symlink to a folder over sftp with and without this patch, with patch it gives error message and no file transfer. Without this patch files from within the folder pointed to are copied, which is wrong. I prefer showing error message over copying the wrong files.