Fix broken iso kioslave
The recent krusader change to not have trailing slash on directory paths exposed a bug in the iso kioslave. One of the code paths did not ensure that the path inside the iso was at least one character ("/"). The other code path also ensured that in case of sub directories, there was no trailing slash... copy that behaviour so that it's the same.