Get default file protocols from KIO instead of hardcoding
Was about to add .tar.xz to the list of hardcoded protocols but decided to change the logic to check which protocol handlers are installed in the system instead.
Now the Protocols konfigurator page is just used for overrides, not for defaults. This way more file formats can be supported just by installing more kio slaves, no need to change krusader code to add support.

Also change how the detection works for the case of user navigating up, outside the archive file. Use the ":local" protocol class to detect generally when user was loooking inside of a file.
Another change is to not "do nothing" in case user tries to enter an archive file which does not have the needed program to decode. Instead, just try it and see if it works. If not then the user will at least see that krusader registered the user's command to open and will then see an error message why it did not work.