Fix broken copy/paste of files
Was passing the list of URLs in the wrong argument to KUrlMimeData::setUrls. Anyway, just skip using that function, we don't need it. Just call the Qt function directly.
Also, there was broken detection of copy/cut from a "virt:/" filesystem. So urls from such places got copied anyway. But that was a good thing... since it seems to work! Removing the broken and unnecessary detection and user error messages.
And remove an unused variable: "destUrl".