Fix crash in KrArc kio slave
It was crashing on exit (destructor) because it had already deleted the krConfig object but was then accessing the KConfigGroup associated with that KConfig object. Could fix by deleting it as the last thing in destructor, but I find it neater to not have it as a separate allocation, just make it part of the kio_krarcProtocol class instead.

There was also a bug in the fullPathName function, the default string was not empty so how could the detectFullPathName function ever get called?

BUG: 349699