Clean up the "bookmark current" action
First, rename it to "add bookmark" as I think that is clearer, using "bookmark" as a verb is not so common. Also no need to say "current" since there is no other action for bookmarking something else. The current location will be default for the new bookmark, yes, but that is normal and expected.. no need to point out.
Then there were two slots, one in KrBookmarkHandler and one in KRslots, for adding a bookmark. I just kept one of them.
Then there were two actions, just kept one of them. This way the active shortcut is displayed in the menu now, user-visible improvement!
Also update the keyboard shortcuts listed in documentation, as I had forgotten to do that before.