Change what happens when executing a file
This rearranges the execute method (which runs when double clicking or pressing enter in a file panel) to first check if there is a registered application to open the file before checking if it is an archive file which can be listed in krusader.
With that change, it is now possible for user to both look inside an archive file (with right arrow key) and open it with external program (enter key). There is no longer a need to disable krusader from opening an archive file.
Also, it makes it possible for krarc to announce support for zip files under cover, like .docx and .odt, without causing a bad user experience.

REVIEW: 124248