Make shortcuts in KrViewer configurable
Adds the actions to the main window's KActionCollection. Adds an action for opening a KShortcutDialog. Removes the "next tab", "previous tab" and "close tab" actions from the menu (to perform these actions.. if you are using the mouse you will click directly on the tab instead of in the menu, if you are using keyboard then most likely you use the shortcuts. Having them in the menu seems pointless then. If you want to see what shortcut to use for an action, you can open the config dialog and see it there. Or read the documentation.)

The rest are just small cleanups (renames, moving things, refactoring a loop), no change in behaviour.

REVIEW: 124247