Make terminals show in foreground
Problem was that any terminals opened after the first one did not appear in the foreground. This change of default command solves the issue, and I can't see any problems with it.

Patch by Alexander Bersenev <bay@hackerdom.ru>. Thanks for the nice help! Keep them coming and let's hope it does not take so long to get applied next time.

REVIEW: 106914