Make use of kr_comparedirs icon
It was removed in previous commit since it was not in use. But then I
noticed that it had been added to breeze icon theme. And I agree that
it is a better icon than the view-split icon.
So, add back the .PNGs (installed as a fallback if user's chosen icon
theme does not have it) and make use of it.