Change many icons used from krusader-specific to system standard icons
Most of the work (the hard part of finding suitable alternative) was done
by Andreas Kainz <kainz.a@gmail.com>. Thank you Andreas!

The only change to his work was to use "user-identity" instead of
"user-properties" as a replacement for "kr_profile". This is because the
identity icon was already used for panel profile selection (in the Window
menu) and I think it is suitable.

In addition to his work I also replaced "zoom-original" with
"document-preview" and "edit" with "document-edit" for the KrViewer.

And then of course remove all the icon files which are no longer used.
The remaining icons installed by krusader will have replacement icons in
the "breeze" and "oxygen" icon set. But they are not in any standard so
any other icon theme is unlikely to have them, we still install them as
fallback icons.

REVIEW: 124958