Solve the problem of Libreoffice not being launched when the user utilizes the "browse archives as directories" mode and double clicks (or presses Enter) on a .odt file, other cases using other kinds of files. Solve a bug searching inside 7z files.
REVIEW: 127369

There was the problem of Krusader not launching e.g. Libreoffice when the user utilized the "browse archives as directories" mode and double clicked (or pressed Enter) on a .ods file (because it was really a .zip file, as it happened in other cases). The function
    // returns true if the right unpacker exist in the system
    KRarcHandler::arcSupported(QString type)
was improved (it has to be improved anyway) and is used so e.g. an .ods file will not be browsed because KRarcHandler::arcSupported() returns false in the case of an .ods file.

Krusader users, in the "browse archives as directories" mode, were also enabled to browse .odt, .docx, .iso, .ods, .xlsx, etc. files using:
    - in the detailed view: the right arrow key, or the "Browse" action available when pressing the secondary button of the mouse.
    - in the brief view: the "Browse" action available when pressing the secondary button of the mouse.
Note: In the "browse archives as directories" mode the forementioned "Browse" action does not appear in the case of e.g. a zip file because there's already an "Open" action that accomplishes that objective.

With those changes one bug was also solved: if e.g. a 7z archive contains a file named "a.txt", the user opens the search tool of Krusader, he specifies that he wants to search for the file "a.txt", he enables the "Search in archives" checkbox, and he clicks on the "Search button" -> now the "a.txt" file is found inside the 7z archive.

Tests done: Browsed zip, rar, tar, tar.bz2, tar.gz, tar.xz, 7z, odt, docx, iso, ods and xlsx files in Krusader, and opened them externally, using combinations of: the detailed view or the brief view, the right arrow key or the primary button of the mouse (in some cases that action launched an external program) or the secondary button of the mouse (its option "Open" or its option "Browse" when it was available), and having the option "browse archives as directories" enabled or disabled.