Make possible that Krusader's embedded terminal doesn't add a lot of `cd` messages to the shell history
REVIEW: 127431

Make possible that Krusader's embedded terminal doesn't add a lot of cd messages to the shell history, that was achieved in Dolphin:
    - https://bugs.kde.org/show_bug.cgi?id=204039
    - https://code.woboq.org/kde/applications/kde-baseapps/dolphin/src/panels/terminal/terminalpanel.cpp.html
    (in that page the string
        m_terminal->sendInput(" cd " + KShell::quoteArg(dir) + '\n');
    can be searched)
and in Krusader the same change was applied.

At least in Ubuntu and similar ones the change really improves the situation, the shell history doesn't get full of cd commands.