Some refactoring: removed classes by merging code + removing code redundancy
Summary:
Yes, code and class diagram style is a matter of taste. But I think removing unnecessary code makes debugging and improvement a lot easier.
Specially because there is not that much active development going on and we want a working application for the next 10 years when porting to Qt7/Kf7 has to be done .)

Commits:

Refactoring: Removed KrInterViewItem class

KrInterViewItem was minimal implementation of abstract class KrViewItem.
Implementation moved to base class.

Refactoring: merged KrItemView with its single subclass KrInterBriefView

+ wheelEvent() in KrItemView was never used.

Simplify and remove duplicate code in Krusader::queryClose()

Refactoring: merged KrMainWindow interface with its single (abstract) subclass FileManagerWindow

Removed FileManagerWindow and moved all declarations to KrMainWindow which is a parent of the Krusader class, so it is already complicated enough.
And krmainwindow.cpp was unused for the last six years.

Test Plan: Used for some days now. Tested file selection in both view types.

Reviewers: #krusader, asensi

Reviewed By: #krusader, asensi

Subscribers: asensi

Differential Revision: https://phabricator.kde.org/D2285