Fix layout of KRMaskChoice
It looks like the "layout" in this dialog was basically a copy of the
generated output of uic, or something like that.

Redo the layout of the dialog from scratch, with the proper creation
order of the widgets, and only the helper sublayouts needed. This also
removes all the hardcoded sizing, and extra class members.