Make KrDebugLogger (and so the e.g. /tmp/krdebug trace file) usable by any file of Krusader.
Interestingly, that also helps when a user has a particular problem that developers can't reproduce. Then a developer can add a lot of `KRFUNC`, `KRDEBUG(...)`, etc. to krusader functions, send the debug Krusader executable to the user, the user can experiment his particular problem, and finally the user is able to send the "krdebug" trace file to the Krusader developers, so they can know better what happened.

Fixes T4999
Differential Revision: http://phabricator.kde.org/D3884