#314258 - Duplicate tab settings when duplicating tabs
Summary: Whenever a new tab is created with this change we'll try to copy configuration from an existing tab. I'm not really familiar with the code but this seems to be the right way to do it.

Test Plan:
When duplicating tabs, various tab settings should be duplicated as well:

* Sort column and order
* Zoom factor
* View type (brief vs. detailed)
* View filter (all files vs. custom)
* "Show Previews" setting

There are also some `ListPanel` properties being duplicated but I don't know how these are set.

Tab duplication should not create unnecessary sections in the config file.

Reviewers: #krusader, abika, martinkostolny, asensi

Reviewed By: #krusader, abika, martinkostolny, asensi

Subscribers: asensi, martinkostolny, abika, #krusader

Tags: #krusader

Differential Revision: https://phabricator.kde.org/D4418

BUG: 314258