#375830 - Don't send SIGINT to terminal emulator unnecessarily
Summary: This leaves clearing current command line as default behavior which is sane. However, this change adds a parameter `clearCommand` to the `sendInput()` method to override this behavior. Inserting the file name and pasting are the scenarios where it is actually undesirable to clear the current command.

Test Plan: Inserting file name (Ctrl+Enter) shouldn't clear command line, both when the focus is inside the terminal and on the file list. Same should go for pasting text (Ctrl+V inside the terminal). The automatic `cd` command sent when current directory changes in the list should still clear command line (bug 328068 should not regress).

Reviewers: #krusader, abika, martinkostolny, asensi

Reviewed By: #krusader, abika, martinkostolny, asensi

Subscribers: asensi, martinkostolny, abika, #krusader

Tags: #krusader

Differential Revision: https://phabricator.kde.org/D4419

BUG: 375830