Panel: Do not select .. in a filtered view if there are other entries
Summary: I commonly find myself in this situation when browsing a directory tree: I filter the view to locate a directory, press Right, then filter again to locate another directory and press Right again. Oops - that directory was the only entry, yet Krusader left .. selected so that I'm now back in the parent directory. While one part of the issue here is the visuals on a high-DPI screen, leaving .. selected when filtering just isn't good for the user experience. In particular, Total Commander seems to select the first entry matching the filter.

Reviewers: martinkostolny, #krusader

Reviewed By: martinkostolny, #krusader

Subscribers: abika, martinkostolny

Differential Revision: https://phabricator.kde.org/D4610