Dialogs: removed support for cfv in checksum dialog
Because cfv is an old tool, is not under active development and does not offer
much advantages over the common tools md5sum, sha1sum, etc. which should be present
on every system.
Additionally this allows much simpler code and will help fixing a security
related bug when verifying checksums.