Panel: setting focus on the panel where New Tab button was clicked
Summary:
FIXED: when pressing a New Tab button the corresponding Panel is not selected

This is a fix for regression introduced years ago (likely even before v2.4).
We have two New Tab buttons on each Panel. Whenever a button is clicked,
a new tab appears, however it was not focused in case
the other Panel was in focus.

Reviewers: #krusader

Differential Revision: https://phabricator.kde.org/D10570