Removed old hack from RadialMap::Map::colorise that is not used nowadays
Whatever it was, I could find no traces of usage of the code and
was not able to hit the code in debugger.
Originating commit dated 2004 and doesn't bring any clarity.
It's safe to remove as the worst that could happen is the color scheme
would be a little off (in case the code still used for something).