Fixed accidentally triggering menus during bookmark search
Because of qt bug 939 setActiveAction triggers menus.
This code ensures that we only highlight non menu items to avoid triggering
the menus until the menu is the only match. It also adds a return true
statement to tell others that we have handled the event that
it should there for not be handled further.

Reviewed By: nmel, martinkostolny
Edited By: nmel

Differential Revision: https://phabricator.kde.org/D11525