Fixed bookmark search key swallowing after exiting from submenu
This patch fixes the following problem:
1. Open bookmark menu.
2. Search for a folder or navigate to it so that submenu is opened.
3. Press Esc to close the submenu.
4. Type a key to search - it start searching but the search is reset
   immediately.

Although #3 sends QCloseEvent event for the submenu, #4 also sends
QCloseEvent event for the submenu (which seems to appear
as we deactivate the folder action) + sends another QCloseEvent event
for the submenu after ~200ms through QTimer for unknown reasons.

This change allows to remember the menu we are currently searching in
and ignore spurious QCloseEvent events.