Fixed active item deselection when starting search after exiting from submenu
This patch fixes the following problem:
1. Open bookmark menu.
2. Search for a folder or navigate to it so that submenu is opened.
3. Press Esc to close the submenu.
4. Type a key to search to find a match - it start searching
   but in a few moments active item becomes deselected.

Although #3 sends QCloseEvent event for the submenu, #4 also sends
QCloseEvent event for the submenu + sends
another QCloseEvent event for the submenu after ~200ms through QTimer
after which the active item is reset.

This change restores the active action in this case through another QTimer.