Fixed regression in "Open in new tab" behaviour
If you are in the folder '/home/klaus', right click the folder 'haus'
(/home/klaus/haus) and choose "open in new tab", then a new tab
should be created with the url /home/klaus/haus.

It was the case in v2.4.0-beta3.
Currently, a new tab will be opened with the url /home/klaus
and the folder haus in focus.
This patch fixes the regression.

Authored By: rade
Reviewed By: nmel, martinkostolny
Edited By: nmel

Differential Revision: https://phabricator.kde.org/D11622