Changed loading of layout.xml from resource system only
layout.xml is a static file, so it's safe to load it from resource.
In the same time, if layout.xml is changed like it happened after v2.6,
this way doesn't introduce problems with loading incorrect layout.
Incorrect layout was loaded if v2.6 is installed through
a package manager and we run locally compiled / installed exe.

FIXED: [ 386314 ] Install to a custom location doesn't work correctly
BUG: 386314

Differential Revision: https://phabricator.kde.org/D12061