Fixed ugly message dialog on Compare by Content error
Change KmessageBox::detailedError dialog type to KmessageBox::sorry for a better view.

When the dialog is created the details are hidden. The detail text is not that long that it couldn't be shown directly.

FIXED: [ 383567 ] Details section is empty in 'Do not know which files to compare.' window
BUG: 383567

Differential Revision: https://phabricator.kde.org/D12904

(cherry picked from commit 94ff5d6ccb3a69627565e345882ae027f475930a)