Panel: Always delete files by their URL
Virtual files could not be deleted because the argument for the delete
operation was the (real) filenames but the filename of a virtual file is
the path to the real file. This is fixed.

FIXED: [ 394896 ] Files fed to listbox (and all other files in virt://) cannot be deleted
BUG: 394896

Differential Revision: https://phabricator.kde.org/D13572

(cherry picked from commit 336fefc02d4a2cd6eeaec73397e3d95552b3766e)