End a quicksearch after an item has been found
When someone uses the search bar to quicksearch for a file: after he presses Return (or Enter) because he has found the file, he still have to press Esc to end the quicksearch. That doesn't happen when using Total Commander, Double Commander or Krusader 2.4.0-beta3, which automatically end those quicksearches.

There are similar use cases: for example when the user quicksearchs for a file, finds it, and (like in Total Commander) presses F4 to edit the file, etc.

Of course, as the KrSearchBar code is being talked about, the improvements of KrSearchBar experts are welcome :-)

FIXED: [ 395285 ] The quicksearch bar is still opened after deleting a file
BUG: 395285

Differential Revision: https://phabricator.kde.org/D13499