Correct the tab order in the Manage User Actions dialog
Note: Trying to move elements (and their labels) in the ui file is difficult and makes difficult to compare versions later. So the "Edit > Edit Tab Order" option of Qt Creator was used to edit the tab order, which is much easier, although in order to achieve that, the process adds a tabstops section at the end of the ui file. That tabstops section is much, much simpler to maintain (using Qt Creator or any text editor), and comparing versions is much saner.

FIXED: [ 396652 ] In the Manage User Actions dialog: the tab order is not the expected one
BUG: 396652

Differential Revision: https://phabricator.kde.org/D14230

(cherry picked from commit 7980ba68014ca901cb90691c9a89e477583b789f)