Prefer to find oktetapart by desktop file, not binary name
Summary:
Okteta >= 0.26 installs the oktetapart binary in the kf5/parts subdir
of the plugins, so the old check will not find it.
But that version also installs a desktop file again, so prefer to use that.

Test Plan:
Okteta KParts plugin is found also with devel version of what will be 0.26
in some weeks.

Reviewers: #krusader, nmel

Reviewed By: #krusader, nmel

Subscribers: nmel, yurchor

Differential Revision: https://phabricator.kde.org/D18881