doc: drop almost all bits related to installation
Summary:
In the majority of the cases, users install krusader as available in
their distribution, and thus most of this content is not useful; users
that want to compile krusader from sources find already instructions on
krusader.org and in the in-source INSTALL files, so no need to
duplicate them in a user-taylored documentation.  Just link to the
online page on krusader.org.

Furthermore:
- instructions related to building distro package are definitely not in
  the scope of an user documentation
- the bits about krusaderui.rc are wrong, since this is is actually a
  bug in krusader (i.e. krusaderui.rc was changed without bumping its
  version)
- links to distro packages (including macOS and Windows ones) either
  are obsolete, or simply not useful
- the reference of the command line options is not useful for general
  users starting krusader graphically; for console users there is
  already the krusader man page with these bits of information (so no
  need to duplicate them)
- the instructions for changing the default file manager are obsolete

Test Plan:
- build the documentation
- verify the mentioned bits are not present

Reviewers: yurchor

Reviewed By: yurchor

Subscribers: yurchor, kde-doc-english

Tags: #documentation

Differential Revision: https://phabricator.kde.org/D20287