Wrap text in copy/move confirmation dialog
When you want to copy/move a file with a very long name, the confirmation dialog is opened very wide to respect the filename width.

Wrapping the text let the dialog to be drawn at the usual width.

FIXED: [ 330788 ] "Information - Krusader" dialog does not fit to screen width; text non-copyable
BUG: 330788

Differential Revision: https://phabricator.kde.org/D22443