Avoid that Krusader crashes when trying to view any archive using the ark part
Note: About checking the return value of the `cpart->openUrl(curl);` sentence, Elvis Angelaccio wrote in <https://bugs.kde.org/show_bug.cgi?id=409722> that "the Ark Part returns false in the openUrl() method even if it is actually opening the archive as requested. While this may be considered a partial abuse of the KParts api, it is done for a reason (that is, ark loads the archives asynchronously). [...] Krusader [could] use the part's completed() and canceled() signals instead."

FIXED: [ 409722 ] Krusader crash when trying to view any archive using the ark part
BUG: 409722

Differential Revision: https://phabricator.kde.org/D23243