Add QT::KeyMenu action to search results
In the Search result list, you can right click for the context menu (view, edit, compare, copy), but the Menu key from the keyboard does not work.

This patch add this functionality.

The popup menu is drawn based on the mouse pointer position.

FIXED: [ 328919 ] Keyboard key "Menu" does not work in list of search results
BUG: 328919

Differential Revision: https://phabricator.kde.org/D23924