Manage what happens when a PanelViewerBase (e.g. using Shift+F3) is being used to open a local file that does not exist
This solves some problems. For example, if users press Shift+F3 and paste the path of a local inexistent file: Krusader opens a viewer window as it that file existed and was empty (but that file is not empty, it really does not exist). Note: That problem doesn't happen in the case of remote inexistent files.

FIXED: [ 395737 ] If the user presses Shift+F3 and writes the name of an inexistent file: Krusader opens a viewer window
BUG: 395737