Refactor the "New Folder..." dialog so that it keeps a folder history and suggests folder names
The "New Folder..." dialog is refactored in order to be able to use a folder
history, which contains a list of folder names that can be reused.

When typing the name of a new folder, if in the folder history there is an entry
that starts the same way as the new name, then folder names are suggested in a
non-intrusive way. Any unwanted element of the folder list can be deleted using
Shift+Del. That folder list is useful, for example, when creating similarly (or
equally) named folders, when creating frequent ones like: incoming, resolved,
accounting, defectives, without_backup, etc.

This commit fixes the "File -> New directory" part of:
    Add separate history for each text field
    https://bugs.kde.org/show_bug.cgi?id=328922

ADDED: The "New Folder..." dialog keeps a manageable folder history and suggests folder names.

Differential Revision: https://phabricator.kde.org/D29316