Konfigurator: Ease management of tooltips and “What's This” texts by changing create* functions. Add a tooltip also useful as a test item
Add inside some `create*` methods the `setWhatsThis()` and `setToolTip()`
callings. That way the management of tooltips and “What's This” help is
facilitated.

Note: The parameter `const QString &toolTip` was already used in:
    KonfiguratorCheckBox *createCheckBox(QString configGroup, QString name, bool defaultValue,
                            QString text, QWidget *parent = nullptr, bool restart = false,
                            const QString &toolTip = QString(), int page = FIRST_PAGE);

A tooltip (related to Krusader lister) was added, with the text "If a text file
is bigger than this size then lister will be used". That tooltip is also useful
as an example and in order to perform tests with it.

GUI: A new translation string was added, which is visible on a tooltip and a “What's This” help.

Differential Revision: https://phabricator.kde.org/D29545