Move 'connect' call before openUrl. Developed by Timur Guzairov
Timur Guzairov ("tguzairov") wrote the main part of the present commit:
Parts::ReadOnlyPart::completed signal is connected after openUrl has been
called, and it won't be emited if openUrl was fast enough. In this case
openUrlFinished won't be emited, which could, for example, cause the text not
receiving a focus.

Differential Revision: https://phabricator.kde.org/D26280