Execute .desktop file
Allows the user to run .desktop instead of opening it in the external editor.

As a .desktop file could lead to security issues (evil Exec=), KIO will warn
the user when he tries to launch it from an unknown location (see
XDG_DATA_DIRS), setting the file with executable bit if the user confirms.

FIXED: [ 394938 ] Can't run .desktop files
BUG: 394938