Fix UserAction icons
The button holding the useraction icon is not set when selecting an element,
but only when the icon is changed by the user. The last saved icon is applied
to the following change if its icon is not changed.

BUG: 328936
BUG: 423434
FIXED: [ 328936 ] "User action manager"; lost icon
FIXED: [ 423434 ] UserAction icon is lost